+++ arrus/core/api/ops/us4r/constraints/TxRxLimits.h
@@ -14,8 +14,8 @@ class TxRxLimits {
 public:
     TxRxLimits(TxLimits tx0, TxLimits tx1, RxLimits rx, const Interval<float> &pri) : tx0(std::move(tx0)), tx1(std::move(tx1)), rx(std::move(rx)), pri(pri) {}
 
-    const TxLimits &getTx0() const { return tx0; }
-    const TxLimits &getTx1() const { return tx1; }
+    const TxLimits &getTxHV0() const { return tx0; }
+    const TxLimits &getTxHV1() const { return tx1; }
     const RxLimits &getRx() const { return rx; }
     const Interval<float> &getPri() const { return pri; }