+++ arrus/core/devices/TxRxParameters.h
@@ -113,7 +113,7 @@ public:
             os << ", center frequency: " << pulse.value().getCenterFrequency()
                << ", n. periods: " << pulse.value().getNPeriods()
                << ", inverse: " << pulse.value().isInverse()
-               << ", tx voltage level: " << pulse.value().getAmplitudeLevel();
+               << ", amplitude: " << pulse.value().getAmplitude();
         }
         else {
             os <<"; tx waveform: ";